requirements.txt
fastapi[all]==0.86.0
uvicorn[standard]==0.18.3
sqlmodel==0.0.8
pandas==1.4.1
scikit-learn<=1.0.2
mlflow==1.26.1
boto3==1.21.22
jupyterlab==3.3.2
pymysql==1.0.2
python-dotenv==0.21.0